카테고리 삭제까지완료

+++ src/main/java/com/javaex/service/BlogService.java
@@ -118,7 +118,18 @@ public class BlogService {
 	//카테고리 삭제
 	public int delCategory(int cateNo) {
 		System.out.println("서비스:카테고리 삭제");
-		return blDao.deleteCate(cateNo);
+		
+		CategoryVo cateVo = blDao.selectByNo(cateNo);
+		int cateCount = cateVo.getCateCount();
+		System.out.println(cateCount);
+		
+		if(cateCount==0) {
+			
+			return blDao.deleteCate(cateNo);
+		}else {
+			return 0;
+		}
+		
 	}
